Improve TimeInput keyboard interaction
