+++ packages/react/.loki/reference/chrome_iphone7_Components_TimeInput_Default.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:228625092230344ce92bc63f149a5028aeb6cc32e897c392f1ecd1f5a3bbc7b3
+size 8190
