+++ packages/react/.loki/reference/chrome_laptop_Components_TimeInput_Default.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:0570407022bfa70bf00bb1e570862ca5fc1e4af5a47757e8051571307cde6723
+size 3110
