+++ packages/react/src/components/timeInput/TimeInput.module.scss
@@ -1,2 +1,71 @@
-.timeInput {
+.input {
+  composes: hds-text-input__input from 'hds-core/lib/components/text-input/text-input.css';
+  display: inline-flex;
+  align-items: center;
+  padding: 0 calc(var(--spacing-s) - 3px);
+
+  &.disabled {
+    background-color: var(--input-background-disabled);
+    border-color: var(--input-border-color-disabled);
+    color: var(--input-color-disabled);
+    cursor: not-allowed;
+
+    &:hover {
+      border-color: var(--input-border-color-disabled);
+    }
+  }
+}
+
+.fullInput {
+  position: absolute;
+  width: 0;
+  height: 0;
+  overflow: hidden;
+  opacity: 0;
+  pointer-events: none;
+}
+
+.partialInput, .divider {
+  border: 0;
+  padding: 0;
+  -webkit-appearance: none;
+  box-sizing: border-box;
+  font-size: 1.125rem;
+  color: var(--input-color-default);
+  line-height: normal;
+  background: transparent;
+}
+
+.partialInput {
+  will-change: transform, box-shadow;
+  padding: 0;
+  text-align: center;
+  width: 1.40rem;
+  overflow: visible;
+
+  &:focus {
+    outline: none;
+    box-shadow: 0 0 0 var(--outline-width) var(--focus-outline-color);
+    transform: translate3d(0, 0, 0);
+
+    &::placeholder {
+      opacity: 0;
+    }
+  }
+}
+
+.divider {
+  width: auto;
+  display: inline-block;
+  margin: 0 5px;
+  pointer-events: none;
+}
+
+.input.disabled {
+  .partialInput, .divider {
+    background-color: var(--input-background-disabled);
+    border-color: var(--input-border-color-disabled);
+    color: var(--input-color-disabled);
+    cursor: not-allowed;
+  }
 }
